Scratch changes to generated artifacts, or casts off artifacts that would be regenerated.

+++ Application/EdFi.Ods.Api/Authentication/ApiClientDetails.cs
@@ -133,7 +133,7 @@ namespace EdFi.Ods.Api.Authentication
             {
                 dto.EducationOrganizationIds = tokenClientRecords
                     .Where(x => x.EducationOrganizationId.HasValue)
-                    .Select(x => x.EducationOrganizationId.Value)
+                    .Select(x => (long) x.EducationOrganizationId.Value)
                     .Distinct()
                     .ToArray();
